requirements.txt
Flask
pandas
openpyxl
msoffcrypto-python
lxml
beautifulsoup4
html5lib